@@ -1,50 +1,53 @@
 {
-    "name": "frontend",
-    "version": "17.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "watch": "ng build --watch --configuration development",
-        "test": "ng test",
-        "lint": "ng lint"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "^17.0.5",
-        "@angular/cdk": "^17.0.2",
-        "@angular/common": "^17.0.5",
-        "@angular/compiler": "^17.0.5",
-        "@angular/core": "^17.0.5",
-        "@angular/forms": "^17.0.5",
-        "@angular/platform-browser": "^17.0.5",
-        "@angular/platform-browser-dynamic": "^17.0.5",
-        "@angular/router": "^17.0.5",
-        "@fullcalendar/angular": "^6.0.3",
-        "@fullcalendar/core": "^6.0.3",
-        "@fullcalendar/daygrid": "^6.0.3",
-        "@fullcalendar/interaction": "^6.0.3",
-        "@fullcalendar/timegrid": "^6.0.3",
-        "chart.js": "^3.3.2",
-        "primeflex": "^3.3.1",
-        "primeicons": "6.0.1",
-        "primeng": "17.2.0",
-        "quill": "^1.3.7",
-        "rxjs": "~7.8.1",
-        "tslib": "^2.5.0",
-        "zone.js": "~0.14.0"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "^17.0.5",
-        "@angular/cli": "^17.0.5",
-        "@angular/compiler-cli": "^17.0.5",
-        "@types/jasmine": "~4.3.1",
-        "jasmine-core": "~4.6.0",
-        "karma": "~6.4.2",
-        "karma-chrome-launcher": "~3.2.0",
-        "karma-coverage": "~2.2.0",
-        "karma-jasmine": "~5.1.0",
-        "karma-jasmine-html-reporter": "~2.1.0",
-        "typescript": "~5.2.2"
-    }
+  "name": "frontend",
+  "version": "17.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test",
+    "lint": "ng lint"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^18.2.11",
+    "@angular/cdk": "^17.0.2",
+    "@angular/common": "^18.2.11",
+    "@angular/compiler": "^18.2.11",
+    "@angular/core": "^18.2.11",
+    "@angular/fire": "^18.0.1",
+    "@angular/forms": "^18.2.11",
+    "@angular/google-maps": "^18.2.12",
+    "@angular/platform-browser": "^18.2.11",
+    "@angular/platform-browser-dynamic": "^18.2.11",
+    "@angular/router": "^18.2.11",
+    "@fullcalendar/angular": "^6.0.3",
+    "@fullcalendar/core": "^6.0.3",
+    "@fullcalendar/daygrid": "^6.0.3",
+    "@fullcalendar/interaction": "^6.0.3",
+    "@fullcalendar/timegrid": "^6.0.3",
+    "chart.js": "^3.3.2",
+    "firebase": "^10.14.1",
+    "primeflex": "^3.3.1",
+    "primeicons": "^6.0.1",
+    "primeng": "17.2.0",
+    "quill": "^1.3.7",
+    "rxjs": "~7.8.1",
+    "tslib": "^2.5.0",
+    "zone.js": "~0.14.10"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^18.2.11",
+    "@angular/cli": "^18.2.11",
+    "@angular/compiler-cli": "^18.2.11",
+    "@types/jasmine": "~4.3.1",
+    "jasmine-core": "~4.6.0",
+    "karma": "~6.4.2",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "~5.5.4"
+  }
 }
