@@ -6,6 +6,12 @@ import { SettingsPageComponent } from './pages/settings-page/settings-page.compo
 import { SocialPageComponent } from './pages/social-page/social-page.component';
 import { LeaderboardComponent } from './subpages/leaderboard/leaderboard.component';
 import { ChatsComponent } from './subpages/chats/chats.component';
+import { InfoPageComponent } from './pages/info-page/info-page.component';
+import { RegisterComponent } from './demo/components/auth/register/register.component';
+import { LoginComponent } from './demo/components/auth/login/login.component';
+import { SpeechPageComponent } from './pages/speech-page/speech-page.component';
+import { TextPageComponent } from './pages/text-page/text-page.component';
+
 
 const routerOptions: ExtraOptions = {
     anchorScrolling: 'enabled'
@@ -13,9 +19,13 @@ const routerOptions: ExtraOptions = {
 
 const routes: Routes = [
     {
-        path: '', component: AppLayoutComponent,
+        path: '',
+        redirectTo: 'auth/register',
+        pathMatch: 'full'
+    },
+    {
+        path: 'user', component: AppLayoutComponent,
         children: [
-
             {
                 path : '',
                 component : DashboardPageComponent,
@@ -42,6 +52,21 @@ const routes: Routes = [
                 path : 'settings',
                 component : SettingsPageComponent,
                 data : { breadcrumb: 'Settings'}
+            },
+            {
+                path : 'info',
+                component : InfoPageComponent,
+                data : { breadcrumb: 'Information'},
+            },
+            {
+                path : 'speech',
+                component : SpeechPageComponent,
+                data : { breadcrumb: 'Speech to Text'}
+            },
+            {
+                path : 'text',
+                component : TextPageComponent,
+                data : { breadcrumb: 'text to speech'}
             },
 
 
