@@ -12,5 +12,5 @@ export interface Product {
     inventoryStatus?: InventoryStatus;
     category?: string;
     image?: string;
-    rating?: number;
+    rank?: number;
 }
